Rename module test_cases to tests
