+++ yare-macro/src/lib.rs
@@ -7,14 +7,14 @@ extern crate proc_macro;
 #[macro_use]
 extern crate syn;
 
-mod test_cases;
+mod tests;
 
 #[proc_macro_attribute]
 pub fn parameterized(
     args: ::proc_macro::TokenStream,
     input: ::proc_macro::TokenStream,
 ) -> ::proc_macro::TokenStream {
-    let test_cases = parse_macro_input!(args as test_cases::TestCases);
+    let test_cases = parse_macro_input!(args as tests::TestCases);
     let test_fn = parse_macro_input!(input as greenhouse::TestFn);
 
     test_cases
